Remove matrix_col_t to allow MATRIX_ROWS > 32 (#10183)
The matrix_col_t type was added in commit 0284431ad9 (part of #3449),
but then the code which used that type was removed in #6140, and no
other users were added since that time.  The presence of that type,
however, limits MATRIX_ROWS to 32, which probably does not matter for a
real keyboard, but prevents doing things like making a firmware to test
all existing pins on a board like Teensy++ 2.0 (which has 46 GPIOs).